+++ source/graphics/2d/Sprites/Object/ObjectSpriteContainer.h
@@ -88,9 +88,9 @@ __CLASS(ObjectSpriteContainer);
 
 void ObjectSpriteContainer_constructor(ObjectSpriteContainer this, u8 spt);
 void ObjectSpriteContainer_destructor(ObjectSpriteContainer this);
-s32 ObjectSpriteContainer_addObjectSprite(ObjectSpriteContainer this, ObjectSprite objectSprite, int numberOfObjects);
-void ObjectSpriteContainer_removeObjectSprite(ObjectSpriteContainer this, ObjectSprite objectSprite, int numberOfObjects);
-bool ObjectSpriteContainer_hasRoomFor(ObjectSpriteContainer this, int numberOfObjects);
+s16 ObjectSpriteContainer_addObjectSprite(ObjectSpriteContainer this, ObjectSprite objectSprite, int numberOfObjects);
+void ObjectSpriteContainer_removeObjectSprite(ObjectSpriteContainer this, ObjectSprite objectSprite, s16 numberOfObjects);
+bool ObjectSpriteContainer_hasRoomFor(ObjectSpriteContainer this, s16 numberOfObjects);
 void ObjectSpriteContainer_setDirection(ObjectSpriteContainer this, int axis, int direction);
 VBVec2D ObjectSpriteContainer_getPosition(ObjectSpriteContainer this);
 void ObjectSpriteContainer_setPosition(ObjectSpriteContainer this, VBVec2D position);
